remove each page body and put provider in top component
