+++ src/App.tsx
@@ -1,6 +1,4 @@
 import React, { Component } from 'react';
-// import CountContainer from './containers/CountContainer';
-// import UsersContainer from './containers/UsersContainer';
 import TopPageContainer from './containers/TopPageContainer';
 import EditPageContainer from './containers/EditPageContainer';
 import NotePage from './components/NotePage';
@@ -8,10 +6,10 @@ import {
   BrowserRouter as Router,
   Route,
 } from 'react-router-dom';
-import { usersStore } from './data_store';
+import { usersStore, noteStore, noteListStore } from './data_store';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import MyAppBar from './components/MyAppBar';
-import TopPageBody from './components/TopPageBody';
+import { Provider } from 'mobx-react';
 
 class App extends Component {
   componentWillMount() {
@@ -22,15 +20,17 @@ class App extends Component {
     return (
       <div>
         <MuiThemeProvider>
-          <Router>
-            <div>
-              <MyAppBar usersStore={usersStore} />
-              <Route exact={true} path="/" component={TopPageContainer} />
-              <Route path="/edit" component={EditPageContainer} />
-              <Route path="/note/:id" component={NotePage} />
-              <Route path="/user/:uid" component={TopPageBody} />
-            </div>
-          </Router>
+          <Provider noteStore={noteStore} usersStore={usersStore} noteListStore={noteListStore}>
+            <Router>
+              <div>
+                <MyAppBar usersStore={usersStore} />
+                <Route exact={true} path="/" component={TopPageContainer} />
+                <Route path="/edit" component={EditPageContainer} />
+                <Route path="/note/:id" component={NotePage} />
+                <Route path="/user/:uid" component={TopPageContainer} />
+              </div>
+            </Router>
+          </Provider>
         </MuiThemeProvider>
       </div>
     );
